+++ Pipfile
@@ -7,6 +7,8 @@ name = "pypi"
 flask = "*"
 flask-sqlalchemy = "*"
 bcrypt = "*"
+flask-migrate = "*"
+flask-script = "*"
 
 [dev-packages]
