+++ server/models.py
@@ -1,4 +1,21 @@
-from app import db
+import bcrypt
+
+from . import db
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(80), unique=True, nullable=False)
+    email = db.Column(db.String(120), unique=True, nullable=False)
+    password_hash = db.Column(db.String(128))
+    is_seller = db.Column(db.Boolean, default=False)
+
+    def set_password(self, password):
+        self.password_hash = bcrypt.hashpw(password.encode(
+            'utf-8'), bcrypt.gensalt()).decode('utf-8')
+
+    def check_password(self, password):
+        return bcrypt.checkpw(password.encode('utf-8'), self.password_hash.encode('utf-8'))
+
 
 class Product(db.Model):
     id = db.Column(db.Integer, primary_key=True)
